code cleanup, generic request methods and more memory management

+++ DragonFruit.Common.Data.Tests/TestBase.cs
@@ -1,14 +1,12 @@
 ﻿// DragonFruit.Common Copyright 2020 DragonFruit Network
 // Licensed under the MIT License. Please refer to the LICENSE file at the root of this project for details
 
-using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using DragonFruit.Common.Data.Tests.BodySerialization;
 using DragonFruit.Common.Data.Tests.FileDownloads;
 using DragonFruit.Common.Data.Tests.SteamNews;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
 namespace DragonFruit.Common.Data.Tests
@@ -21,25 +19,50 @@ namespace DragonFruit.Common.Data.Tests
         [TestMethod]
         public void TestJsonRequest()
         {
-            Client.Perform<SteamNewsResponse>(new SteamNewsRequest());
+            var request = new SteamNewsRequest();
+
+            //returns the data
+            Client.Perform<SteamNewsResponse>(request);
+
+            //returns just the response info - i.e a head request where serialization may not be desired
+            Assert.IsTrue(Client.Perform(request).IsSuccessStatusCode);
         }
 
         [TestMethod]
         public void TestJsonBodyRequest()
         {
             var request = new DatabaseUpdateRequest();
-            var response = Client.Perform<JObject>(request);
 
-            Assert.AreEqual(request.Employee.Department, response["json"].ToObject<Employee>().Department);
+            var firstResponse = Client.Perform<JObject>(request);
+            Assert.AreEqual(request.Employee.Department, firstResponse["json"].ToObject<Employee>().Department);
+
+            var secondResponse = Client.PerformLast<JObject>();
+            Assert.AreEqual(secondResponse["json"].ToObject<Employee>().Department, firstResponse["json"].ToObject<Employee>().Department);
         }
 
         [TestMethod]
         public void TestFileRequest()
         {
             var request = new FileDownloadRequest();
-            Client.Perform(request);
 
-            File.Delete(request.Destination);
+            Assert.IsFalse(File.Exists(request.Destination));
+
+            try
+            {
+                Client.Perform(request);
+
+                Assert.IsTrue(File.Exists(request.Destination));
+                Assert.IsTrue(new FileInfo(request.Destination).Length > 5000);
+
+                File.Delete(request.Destination);
+            }
+            catch
+            {
+                if (File.Exists(request.Destination))
+                    File.Delete(request.Destination);
+
+                throw;
+            }
         }
     }
 }
